Round message sending frequency to integer (floating point numbers are meaningless) * Also made some GUI adjustments

+++ app/src/main/java/app/mordred/whatscloud/model/UserListItem.kt
@@ -2,4 +2,4 @@ package app.mordred.whatscloud.model
 
 import android.graphics.Bitmap
 
-data class UserListItem(val usrName: String, val usrMsgCount: Int, val usrMsgFreq: Float, val usrWordCloud: Bitmap)
+data class UserListItem(val usrName: String, val usrMsgCount: Int, val usrMsgFreq: Int, val usrWordCloud: Bitmap)
